💥 Return instances of ImageData for bmp images
Migration Guide:

When returning bmp images, this library will not return instances of `ImageData`. The only change that this should bring is that the `data` property is now an `Uint8ClampedArray` instead of `Uint8Array` when the `type` property is equal to `"bmp"`.

The `data` property will still be an `Uint8Array` if the returned `type` is equal to `"png"`.